Use transactions in AccessAPI
Without transactions in this API, in the old way it was created, access
could be lost as access was first deleted, and then access objects were
created, and finally new access was added. If any of the two last steps
failed, all access would still be lost.

This avoids both the problem described above, as well as possible race
conditions.